Move add authentication server dialog to a new file
